review: harden store caps per CodeRabbit feedback on #66
- config: drop (int) casts on store_caps env reads; capValue() now parses and
  validates, falling back to the default on a malformed value instead of
  silently coercing to 0 (which would disable the very cap meant to bound
  memory). A literal 0 still disables on purpose.
- CoverageRegistry: maintain a running entry counter so the per-record cap
  check is O(1) instead of rescanning buckets; entryCount() now reads it.
- IdentityMapStore::remember(): promoting an existing absent marker to a live
  entry is net-zero, so it no longer trips the cap (was an over-eager flush).
- tests: absent->entry promotion at full budget does not flush; unique-key
  index flushes on its own cap without touching the entry store; capValue()
  semantics (malformed->default, 0->disabled, numeric string/int parsed).

+++ config/query-ricer-extreme.php
@@ -101,11 +101,15 @@ return [
      *   max_entries          — IdentityMapStore $entries + $absent combined.
      *   max_unique_keys      — UniqueKeyIndex live + absent fingerprints.
      *   max_coverage_entries — CoverageRegistry recorded regions.
+     *
+     * Values are parsed (and validated) in the service provider rather than cast
+     * here, so a malformed env value falls back to the safe default instead of
+     * silently coercing to 0 and disabling the cap. Use a literal 0 to disable.
      */
     'store_caps' => [
-        'max_entries' => (int) env('IDENTITY_MAP_MAX_ENTRIES', 100000),
-        'max_unique_keys' => (int) env('IDENTITY_MAP_MAX_UNIQUE_KEYS', 100000),
-        'max_coverage_entries' => (int) env('IDENTITY_MAP_MAX_COVERAGE_ENTRIES', 50000),
+        'max_entries' => env('IDENTITY_MAP_MAX_ENTRIES', 100000),
+        'max_unique_keys' => env('IDENTITY_MAP_MAX_UNIQUE_KEYS', 100000),
+        'max_coverage_entries' => env('IDENTITY_MAP_MAX_COVERAGE_ENTRIES', 50000),
     ],
 
     /*
